Revert "Submit gist in popup, not background page"
This reverts commit caf6882c739adc3266a6b6274f9a3d25315ccc84.

Instead of accomodating the lack of activeTab permissions in the
background page by moving the relevant code to a place where
activeTab is respected (the popup), we should just eschew activeTab
as the mechanism to grant access to the page, as when this code
lives in the popup, it's subject to premature termination if the
popup is closed before the Gist is fully submitted. As such, this
commit replaces the "activeTab" permission with the broader
"<all_urls>" permission, which, while slightly less elegant, will
give the correct behavior.

Hopefully, at some point in the future, crbug.com/695340 will be
fixed, and this change can be fully reverted to its original form
(using both the background page *and* "activeTab" permission).